requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
bcrypt
pydantic
